Move upstream to root
