--- internal/http/upstream/doc.go
@@ -1,4 +1,0 @@
-/*
-Package upstream offers different upstream implementations used as http.Handler
-*/
-package upstream
