--- internal/http/upstream/reverseproxy.go
@@ -1,1 +1,0 @@
-package upstream
