+++ upstream/doc.go
@@ -1,0 +1,2 @@
+// Package upstream offers different upstream implementations used as http.Handler
+package upstream
